Various Win32 fixes. Change args in do_ms.bat to put platform last. Fix unsigned/signed cmp error in asn1parse. Change various pem_all.c args to use pem_password_cb.

+++ apps/asn1pars.c
@@ -281,7 +281,7 @@ bad:
 
 	if (length == 0) length=(unsigned int)num;
 	if(derout) {
-		if(BIO_write(derout, str + offset, length) != length) {
+		if(BIO_write(derout, str + offset, length) != (int)length) {
 			BIO_printf(bio_err, "Error writing output\n");
 			ERR_print_errors(bio_err);
 			goto end;
